Move (models, views, controllers) in src
